+++ e2e/tests-app-ng/app/main.ts
@@ -1,8 +1,8 @@
-import { platformNativeScript } from "@nativescript/angular";
+import { platformNativeScriptDynamic } from "@nativescript/angular";
 
 import { AppModule } from "./app.module";
 import { Trace } from "@nativescript/core";
 
 Trace.enable();
 
-platformNativeScript().bootstrapModule(AppModule);
+platformNativeScriptDynamic().bootstrapModule(AppModule);
