+++ nativescript-angular-package/common.ts
@@ -1,1 +1,1 @@
-export * from "@nativescript/angular/common";
+export * from "@nativescript/angular";
