+++ nativescript-angular-package/forms/index.ts
@@ -1,1 +1,1 @@
-export * from "@nativescript/angular/forms"
+export * from "@nativescript/angular";
