+++ nativescript-angular-package/index.ts
@@ -1,9 +1,1 @@
 export * from "@nativescript/angular";
-export * from "@nativescript/angular/forms";
-export * from "@nativescript/angular/router";
-export * from "@nativescript/angular/file-system/ns-file-system";
-export * from "@nativescript/angular/modal-dialog";
-export * from "@nativescript/angular/forms/value-accessors/base-value-accessor";
-export * from "@nativescript/angular/trace";
-export * from "@nativescript/angular/renderer";
-export * from "@nativescript/angular/view-util";
