+++ nativescript-angular-package/platform-providers.ts
@@ -1,1 +1,1 @@
-export * from "@nativescript/angular/platform-providers";
+export * from "@nativescript/angular";
