+++ nativescript-angular-package/router/index.ts
@@ -1,1 +1,1 @@
-export * from "@nativescript/angular/router";
+export * from "@nativescript/angular";
