+++ nativescript-angular-package/router/router.module.ts
@@ -1,2 +1,1 @@
-export { LocationState } from "@nativescript/angular/router/ns-location-strategy";
 export * from "@nativescript/angular/router/router.module";
