+++ nativescript-angular-package/testing/src/nativescript_test_component_renderer.ts
@@ -1,1 +1,1 @@
-export * from "@nativescript/angular/testing/src/nativescript_test_component_renderer";
+export * from "@nativescript/angular/testing";
