+++ nativescript-angular-package/value-accessors/base-value-accessor.ts
@@ -1,3 +1,2 @@
 // This file is only for compatibility with pre 4.4.0 releases.
-// Please use "nativescript-angular/forms/value-accessors/base-value-accessor"
-export * from "@nativescript/angular/forms/value-accessors/base-value-accessor";
+export * from "@nativescript/angular";
