+++ nativescript-angular/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@nativescript/angular",
-  "version": "10.0.0-alpha.0",
+  "version": "10.0.0-rc.4",
   "description": "An Angular renderer that lets you build mobile apps with NativeScript.",
   "homepage": "https://www.nativescript.org/",
   "bugs": "https://github.com/NativeScript/nativescript-angular/issues",
@@ -58,6 +58,7 @@
   },
   "dependencies": {
     "@nativescript/zone-js": "~1.0.0",
+    "nativescript-angular": "rc",
     "nativescript-intl": "^4.0.0"
   },
   "peerDependencies": {
@@ -79,7 +80,7 @@
     "@angular/platform-browser": "~10.0.0",
     "@angular/platform-browser-dynamic": "~10.0.0",
     "@angular/router": "~10.0.0",
-    "@nativescript/core": "~7.0.0-rc.12",
+    "@nativescript/core": "rc",
     "codelyzer": "^5.2.0",
     "conventional-changelog-cli": "^2.0.34",
     "husky": "^4.2.5",
