+++ nativescript-angular/platform.ts
@@ -5,7 +5,7 @@ import { NativeScriptPlatformRef, AppOptions, PlatformFactory, COMMON_PROVIDERS 
 // "Static" platform
 const _platformNativeScript: PlatformFactory = createPlatformFactory(platformCore, 'nativeScript', [...COMMON_PROVIDERS]);
 
-export function platformNativeScript(options?: AppOptions, extraProviders?: any[]): PlatformRef {
+export function platformNativeScriptDynamic(options?: AppOptions, extraProviders?: any[]): PlatformRef {
 	// Return raw platform to advanced users only if explicitly requested
 	if (options && options.bootInExistingPage === true) {
 		return _platformNativeScript(extraProviders);
