+++ nativescript-angular/renderer.ts
@@ -7,77 +7,12 @@ import { ViewUtil } from './view-util';
 import { NgView, InvisibleNode } from './element-registry';
 import { NativeScriptDebug } from './trace';
 
-// CONTENT_ATTR not exported from NativeScript_renderer - we need it for styles application.
-const COMPONENT_REGEX = /%COMP%/g;
-export const COMPONENT_VARIABLE = '%COMP%';
-export const HOST_ATTR = `_nghost-${COMPONENT_VARIABLE}`;
-export const CONTENT_ATTR = `_ngcontent-${COMPONENT_VARIABLE}`;
-const ATTR_SANITIZER = /-/g;
-
 export interface ElementReference {
 	previous: NgView;
 	next: NgView;
 }
 
 @Injectable()
-export class NativeScriptRendererFactory implements RendererFactory2 {
-	private componentRenderers = new Map<string, NativeScriptRenderer>();
-	private viewUtil: ViewUtil;
-	private defaultRenderer: NativeScriptRenderer;
-	private rootNgView: NgView;
-
-	constructor(@Optional() @Inject(APP_ROOT_VIEW) rootView: View, @Inject(DEVICE) device: IDevice, private zone: NgZone) {
-		this.viewUtil = new ViewUtil(device);
-		this.setRootNgView(rootView);
-		this.defaultRenderer = new NativeScriptRenderer(this.rootNgView, zone, this.viewUtil);
-	}
-
-	private setRootNgView(rootView: any) {
-		if (!rootView) {
-			rootView = getRootPage();
-		}
-
-		rootView.nodeName = 'NONE';
-		this.rootNgView = rootView;
-	}
-
-	createRenderer(element: any, type: RendererType2): NativeScriptRenderer {
-		if (!element || !type) {
-			return this.defaultRenderer;
-		}
-
-		let renderer = this.componentRenderers.get(type.id);
-		if (renderer) {
-			if (renderer instanceof EmulatedRenderer) {
-				renderer.applyToHost(element);
-			}
-
-			return renderer;
-		}
-
-		if (type.encapsulation === ViewEncapsulation.None) {
-			type.styles.map((s) => s.toString()).forEach(addStyleToCss);
-			renderer = this.defaultRenderer;
-		} else {
-			renderer = new EmulatedRenderer(type, this.rootNgView, this.zone, this.viewUtil);
-			(<EmulatedRenderer>renderer).applyToHost(element);
-		}
-
-		this.componentRenderers.set(type.id, renderer);
-		return renderer;
-	}
-
-	ngOnDestroy(): void {
-		if (NativeScriptDebug.isLogEnabled()) {
-			NativeScriptDebug.rendererLog(`NativeScriptRendererFactory.ngOnDestroy()`);
-		}
-
-		while (this.rootNgView && this.rootNgView.firstChild) {
-			this.viewUtil.removeChild(this.rootNgView, this.rootNgView.firstChild);
-		}
-	}
-}
-
 export class NativeScriptRenderer extends Renderer2 {
 	data: { [key: string]: any } = Object.create(null);
 
@@ -297,58 +232,4 @@ export class NativeScriptRenderer extends Renderer2 {
 		}
 		return () => renderElement.off(eventName, zonedCallback);
 	}
-}
-
-class EmulatedRenderer extends NativeScriptRenderer {
-	private contentAttr: string;
-	private hostAttr: string;
-
-	constructor(component: RendererType2, rootView: NgView, zone: NgZone, viewUtil: ViewUtil) {
-		super(rootView, zone, viewUtil);
-
-		const componentId = component.id.replace(ATTR_SANITIZER, '_');
-		this.contentAttr = replaceNgAttribute(CONTENT_ATTR, componentId);
-		this.hostAttr = replaceNgAttribute(HOST_ATTR, componentId);
-		this.addStyles(component.styles, componentId);
-	}
-
-	applyToHost(view: NgView) {
-		super.setAttribute(view, this.hostAttr, '');
-	}
-
-	appendChild(parent: any, newChild: NgView): void {
-		super.appendChild(parent, newChild);
-	}
-
-	createElement(parent: any, name: string): NgView {
-		const view = super.createElement(parent, name);
-
-		// Set an attribute to the view to scope component-specific css.
-		// The property name is pre-generated by Angular.
-		super.setAttribute(view, this.contentAttr, '');
-
-		return view;
-	}
-
-	@profile
-	private addStyles(styles: (string | any[])[], componentId: string) {
-		styles
-			.map((s) => s.toString())
-			.map((s) => replaceNgAttribute(s, componentId))
-			.forEach(addScopedStyleToCss);
-	}
-}
-
-// tslint:disable-next-line
-const addStyleToCss = profile('"renderer".addStyleToCss', function addStyleToCss(style: string): void {
-	Application.addCss(style);
-});
-
-// tslint:disable-next-line
-const addScopedStyleToCss = profile('"renderer".addScopedStyleToCss', function addScopedStyleToCss(style: string): void {
-	Application.addCss(style, true);
-});
-
-function replaceNgAttribute(input: string, componentId: string): string {
-	return input.replace(COMPONENT_REGEX, componentId);
 }
