+++ nativescript-angular/testing/src/public_api.ts
@@ -1,0 +1,5 @@
+export * from './util';
+export * from './test-root-view';
+export * from './nativescript_test_component_renderer';
+export * from './nativescript-testing.module';
+
